chore: move halstead utilities to utils folder
